+++ changelog.md
@@ -1,12 +1,12 @@
 ### Changes and new features
 
-- 
+- A new option to display the original image (without color adaptation) as a preview;
 
 <details>
 <summary>Fixes and minor changes</summary>
 
 - The Valley staircase algorithm has been rewritten and optimized;
 - Fixed reading air and unexisting blocks from complete_palette.json;
-- Fixed mouse scroll behavior in Presets Editor above and below the window box;
+- Fixed mouse scroll behavior in the preset editor above and below the window box;
 
 </details>